feat: move relations block data to own hook
